Rename functions folder to utils
Renaming from something completely inappropriate. I think utils works.
It's at least consistent with what we have in the react project. Let's
see how it works out.
