+++ packages/harpstrata/src/harpstrata/IsActive/utils/getCounterpartActiveIds/index.test.ts
